git_automation/create.py
import sys
import os
from github import Github
from dotenv import load_dotenv

load_dotenv()

path = os.getenv("FILEPATH")
username = os.getenv("USERNAME")
password = os.getenv("PASSWORD")

def create():
    folderName = str(sys.argv[1])
    folderpath = os.path.join(path,folderName)
    if os.path.exists(folderpath):
        print("Folder already exists.. Link to the path - "+ folderpath)
    os.makedirs(folderpath)
    user = Github(username, password).get_user()
    repo = user.create_repo(sys.argv[1])
    print("Succesfully created repository {}".format(sys.argv[1]))


if __name__ == "__main__":
    create()
