git_automation/requirements.txt
selenium
PyGithub
python-dotenv
